pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-drf-mcp"
version = "1.0.0"
description = "Django REST Framework with Model Context Protocol integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Joel Hanson", email = "joelhanson025@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0.8",
    "djangorestframework>=3.15.2",
    "mcp>=1.12.3",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "black",
    "isort",
    "flake8",
]

[project.urls]
Homepage = "https://github.com/joelhanson/django-drf-mcp"
Repository = "https://github.com/joelhanson/django-drf-mcp.git"
Documentation = "https://github.com/joelhanson/django-drf-mcp#readme"
"Bug Tracker" = "https://github.com/joelhanson/django-drf-mcp/issues"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
skip_glob = ["*/migrations/*"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
]
